// src/app/Models/task.ts
export class Task {
  Lid: any;
  Task_Description: any;
  Task_Comment: any;
  // tslint:disable-next-line: variable-name
  Task_Status_id: any;
  // tslint:disable-next-line: variable-name
  Task_Duedate: any;
  Task_Priority: any;
  Task_Name: any;
  Task_Assigned: any;
  // tslint:disable-next-line: variable-name
  Task_Created: any;
  Task_Updated: any;
}